test: add "Should do completion"
