+++ vscode-styled-jsx-languageserver/client/src/test/completion.test.ts
@@ -1,61 +1,46 @@
 import * as vscode from 'vscode'
 import * as assert from 'assert'
-import { getDocUri, activate } from './helper'
+import { getDocUri, testCompletion } from './helper'
 
 suite('Should do completion', () => {
-  const docUri = getDocUri('highlight.js')
+  const highlightUri = getDocUri('highlight.js')
+  const multiLineUri = getDocUri('multi-line.tsx')
+  const normalUri = getDocUri('normal.jsx')
 
   test('Completes CSS in styled-jsx component', async () => {
-    await testCompletion(docUri, new vscode.Position(7, 11), {
-      items: [
-        { label: 'blue', kind: vscode.CompletionItemKind.Color },
-        { label: 'red', kind: vscode.CompletionItemKind.Color },
-        { label: 'yellow', kind: vscode.CompletionItemKind.Color }
-      ]
-    })
+    assert.strictEqual(
+      await testCompletion(highlightUri, new vscode.Position(7, 11), {
+        items: [
+          { label: 'blue', kind: vscode.CompletionItemKind.Color },
+          { label: 'red', kind: vscode.CompletionItemKind.Color },
+          { label: 'yellow', kind: vscode.CompletionItemKind.Color }
+        ]
+      }),
+      true
+    )
 
-    await testCompletion(docUri, new vscode.Position(8, 12), {
-      items: [
-        { label: 'block', kind: vscode.CompletionItemKind.Value },
-        { label: 'inline', kind: vscode.CompletionItemKind.Value },
-        { label: 'inline-block', kind: vscode.CompletionItemKind.Value },
-        { label: 'flex', kind: vscode.CompletionItemKind.Value },
-        { label: 'grid', kind: vscode.CompletionItemKind.Value }
-      ]
-    })
+    assert.strictEqual(
+      await testCompletion(multiLineUri, new vscode.Position(14, 18), {
+        items: [
+          { label: 'block', kind: vscode.CompletionItemKind.Value },
+          { label: 'inline', kind: vscode.CompletionItemKind.Value },
+          { label: 'inline-block', kind: vscode.CompletionItemKind.Value },
+          { label: 'flex', kind: vscode.CompletionItemKind.Value },
+          { label: 'grid', kind: vscode.CompletionItemKind.Value }
+        ]
+      }),
+      true
+    )
+
+    assert.strictEqual(
+      await testCompletion(normalUri, new vscode.Position(21, 19), {
+        items: [
+          { label: 'absolute', kind: vscode.CompletionItemKind.Value },
+          { label: 'relative', kind: vscode.CompletionItemKind.Value },
+          { label: 'fixed', kind: vscode.CompletionItemKind.Value }
+        ]
+      }),
+      true
+    )
   })
 })
-
-function includes (
-  actualArray: vscode.CompletionItem[],
-  expectedArray: vscode.CompletionItem[]
-): boolean {
-  const result: boolean[] = []
-  expectedArray.forEach(item => {
-    const hasItem = actualArray.some(item2 => {
-      return item2.kind === item.kind && item2.label === item.label
-    })
-    if (hasItem) {
-      result.push(true)
-    }
-  })
-  return expectedArray.length === result.length
-}
-
-async function testCompletion (
-  docUri: vscode.Uri,
-  position: vscode.Position,
-  expectedCompletionList: vscode.CompletionList
-): Promise<void> {
-  await activate(docUri)
-  const actualCompletionList: vscode.CompletionList = await vscode.commands.executeCommand(
-    'vscode.executeCompletionItemProvider',
-    docUri,
-    position
-  )
-  const actualArrayIncludesExpectedArray = includes(
-    actualCompletionList.items,
-    expectedCompletionList.items
-  )
-  assert.strictEqual(actualArrayIncludesExpectedArray, true)
-}
