+++ vscode-styled-jsx-languageserver/client/testFixture/multi-line.tsx
@@ -12,6 +12,7 @@ const Header = () => (
       <h1>Header</h1>
       <style jsx>{`
         .root {
+          display: ;
           max-width: 760px;
           padding-top: 5px;
         }
